Refactor IOCDialog
